+++ package.json
@@ -40,6 +40,7 @@
     "file-loader": "^1.1.11",
     "html-webpack-plugin": "^3.2.0",
     "node-sass": "^4.9.0",
+    "offline-plugin": "^5.0.5",
     "postcss-loader": "^2.1.5",
     "resolve-url-loader": "^2.3.0",
     "sass-loader": "^7.0.1",
