+++ src/balance.js
@@ -1,0 +1,4 @@
+export const initializeBalance = balanceData => {
+    const balanceSpan = document.querySelector('.bitcoin--balance');
+    balanceSpan.textContent = balanceData.confirmed_balance;
+}
