+++ webpack.config.js
@@ -1,5 +1,7 @@
 const webpack = require('webpack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
+const OfflinePlugin = require('offline-plugin')
+
 const { author, config , version } = require('./package.json');
 const { address } = require('./btc-config.json');
 
@@ -47,6 +49,15 @@ module.exports = {
             title: config.title,
             version: `v${version}`,
             inject: false
+        }),
+        new OfflinePlugin({
+            version: '[hash]',
+            responseStrategy: 'network-first',
+            safeToUseOptionalCaches: true,
+            ServiceWorker: {
+                events: true,
+                output: 'static/js/sw.js'
+            }
         })
     ],
     devtool: 'eval-source-map',
